Test_training of AI
the reward function is ass and from chatbot. This is the most important thing to change.
I worked a lot on the env_base to make it work and also changed our main so that the environment is being used to run the simulation and not just mujoco. We will not need the mallet.xml anymore so that will be removed in the future. I also changed the observations of the environment